% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks_rules.R
\name{breaks_rules}
\alias{breaks_rules}
\title{Sets the rules to categorize frontier metrics}
\usage{
breaks_rules(
  baseline = list("jenks", c("Low", "Medium", "High")),
  baseline_frag = list("jenks", c("Low", "Medium", "High")),
  loss = list("jenks", c("Low", "Medium", "High")),
  loss_frag = list("jenks", c("Low", "Medium", "High")),
  speed = list("jenks", c("Slow", "Medium", "Fast")),
  left = list("jenks", c("Low", "Medium", "High"))
)
}
\arguments{
\item{baseline, baseline_frag, loss, loss_frag, speed, left}{A list defining the
rules to define discrete classes for each frontier metric. See Details.}
}
\value{
An object of class 'FronterMetric_classes' to be used in argument \code{breaks} within \code{\link[=fmetrics]{fmetrics()}}.
}
\description{
Defines rules to generate discrete categories for continuous frontier metrics.
}
\details{
For each metric, the user must provide a list. The first element of the list
defines the algorithm to be used for the categorization. The second element of  the list must be a character vector with the names
for each category that will be generated.

The following algoritms are available (first element of the list):

(1) "jenks" (default): Jenks natural breaks classification (uses BAMMtools::getJenksBreaks).
The number (n) of intervals is given by the number of categories provided
in the second element of the list.

(2) "equal": splits the numeric range \link{min(y), max(y)} of metric values (\code{y}) into n intervals of equal
width. The number (n) of intervals is given by the number of categories provided
in the second element of the list.

(3) "quantile": splits observations by quantiles so that each bin contains (roughly)
the same number of observations (frontier cells). Quantiles are calculated given
the number of categories provided in the second element of the list. For instance,
if 4 categories are provided, the following intervals be used as cuts: 0-0.25, 0.25-0.5, 0.5-0.75, and 0.75-1.

(4) Alternatively, a numeric vector with two or more unique cut points can be
provided. For example: \code{c(5, 10, 55, Inf)}, and the categories (second element
of the list) could be \code{c("Low", "Medium", "High")}, where "Low": 5-10, "Medium": 10-55, and
"High": 55-Inf.

To define new rules, the user must provide a list for each frontier metric. For all metrics except "activeness"
the rules should follow the format used in \code{\link[=cut]{cut()}}. Specifically, the first element
of the list should be a numeric vector of two or more unique cut points (breaks), and the second
element should be a character vector of labels corresponding to each resulting category. For
instance, classes for frontier metric "baseline" could be ruled as follows:

When assigned to an object, the new definition of classes can be passed to \code{\link[=fmetrics]{fmetrics()}},
inside the argument \code{breaks}, to generate frontier metrics's classes with a new definition.
}
\examples{
\dontrun{
# Explore default parameters
init_classes()

# Define new parameters for baseline and activeness
# Baseline
#   Low: between 5 and 20
#   Medium: between 20 and 70
#   High: Higher than 70
# Activeness
#   Old frontiers: last year active in 2019
#   Emerging frontiers: first year active in 2020
class_par <- init_classes(baseline = list(c(5, 20, 70, Inf),
                                          c("Low", "Medium", "High")),
                          activeness = list(old = 2019, emerging = 2020, window = 5))

# Shows new defined classes
class_par

# This object can be passes to [fmetrics()] inside the argument
# `classes`, to generate frontier metrics's classes with this new definition
}
}
