% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{copo_metrics}
\alias{copo_metrics}
\title{'FrontierMetric' object}
\format{
An object of class \code{FrontierMetric} of length 1.
}
\usage{
copo_metrics
}
\description{
An object of class 'FrontierMetric' generated with \code{\link[=fmetrics]{fmetrics()}} for its use in examples.
It contains the values of all available frontier metrics for the surroundings of the
Copo National Park (Argentina).
}
\details{
The main internal object is a data frame (accesible through \code{copo_metrics@data}) containing all
metric values for this study area.
}
\seealso{
See the examples sections of \code{\link[=fmetrics]{fmetrics()}}, \code{\link[=fmetrics_plot]{fmetrics_plot()}}, \code{\link[=fmetrics_summary]{fmetrics_summary()}} and \code{\link[=fmetrics_rast]{fmetrics_rast()}}
for more context.
}
\keyword{datasets}
