% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmetrics_summary.R
\name{fmetrics_summary}
\alias{fmetrics_summary}
\title{Calculates summary statistics of frontier metrics}
\usage{
fmetrics_summary(x, metrics = "all", show_plots = TRUE)
}
\arguments{
\item{x}{Object of class 'FrontierMetric' generated with \code{\link[=fmetrics]{fmetrics()}}.}

\item{metrics}{Frontier metrics to consider. Options include one or more metrics contained in \code{x},
or "all" to calculate all frontier metrics available within \code{x}. Default is "all".}

\item{show_plots}{Logical. If \code{TRUE} (default), histograms will be plotted when
running the function.}
}
\value{
A list with three elements:
(1) a data frame with basic summary statistics;
(2) a data frame with the total area of each class within each frontier metric; and
(3) a ggplot object with histograms of frontier metrics.
}
\description{
Calculates statistics of continuous values of frontier metrics, plots
histograms, and calculates the total area of each discrete class of frontier metrics.
}
\details{
Access to statistics once generated the object can be done with \verb{@summary_stats} and \verb{@classes_areas}.
To access and plot the histograms, use \verb{@hists}.

User-defined metrics will not be considered.
}
\examples{
summ <- fmetrics_summary(copo_metrics)
summ

# Access to summary statistics
summ@summary_stats

# Access to classes areas
summ@classes_areas

# Plotting of histograms
summ@hists
}
