% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{FrontierMetric-class}
\alias{FrontierMetric-class}
\title{Class 'FrontierMetric'}
\description{
Object of class 'FrontierMetric' are generated with \code{\link[=fmetrics]{fmetrics()}}, and hold
the values of the calculated frontier metrics for each cell of the study area.
Objects of this class can be passed to \code{\link[=fmetrics_plot]{fmetrics_plot()}} and \code{\link[=fmetrics_rast]{fmetrics_rast()}}.
}
\section{Slots}{

\describe{
\item{\code{metrics}}{Calculated frontier metrics.}

\item{\code{time_frame}}{Year range of the studied time series.}

\item{\code{data}}{Main data frame containing the values of the calculated frontier
metrics for each individual cell of the study area.}

\item{\code{extent}}{Extent of the study area.}

\item{\code{grain}}{Smallest and biggest studied grain (resolution).}

\item{\code{aggregation}}{Pre-defined aggregation during running of \code{\link[=init_fmetrics]{init_fmetrics()}}.}

\item{\code{min_treecover}}{Pre-defined minimum tree cover during running of \code{\link[=init_fmetrics]{init_fmetrics()}}.}

\item{\code{min_cover}}{Pre-defined minimum forest cover during running of \code{\link[=init_fmetrics]{init_fmetrics()}}.}

\item{\code{min_rate}}{Pre-defined minimum forest loss rate during running of \code{\link[=init_fmetrics]{init_fmetrics()}}.}

\item{\code{window}}{Pre-defined window of years to calculate \code{min_rate} during running of \code{\link[=init_fmetrics]{init_fmetrics()}}.}

\item{\code{excluded_cells}}{Cells excluded for future analysis (not frontiers)}
}}

