% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{FrontierMetric_breaks-class}
\alias{FrontierMetric_breaks-class}
\title{Class 'FrontierMetric_breaks'}
\description{
Objects of class 'FrontierMetric_breaks' are generated with \code{\link[=breaks_rules]{breaks_rules()}}.
It holds the definition of rules to categorize the values of frontier metrics
into discrete classes.
}
\section{Slots}{

\describe{
\item{\code{baseline}}{List of rules for frontier metric "baseline forest".}

\item{\code{baseline_frag}}{List of rules for frontier metric "baseline forest fragmentation".}

\item{\code{loss}}{List of rules for frontier metric "forest loss".}

\item{\code{loss_frag}}{List of rules for frontier metric "forest loss fragmentation".}

\item{\code{speed}}{List of rules for frontier metric "speed".}

\item{\code{left}}{List of rules for frontier metric "forest left".}
}}

