% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{FrontierMetric_classes-class}
\alias{FrontierMetric_classes-class}
\title{Class 'FrontierMetric_classes'}
\description{
Objects of class 'FrontierMetric_classes' are generated with \code{\link[=init_classes]{init_classes()}}.
It holds the definition of rules to categorize the values of frontier metrics
into discrete classes.
}
\section{Slots}{

\describe{
\item{\code{baseline}}{List of rules for frontier metric " baseline woodland".}

\item{\code{loss}}{List of rules for frontier metric "woodland loss".}

\item{\code{fragmentation}}{List of rules for frontier metric "fragmentation".}

\item{\code{speed}}{List of rules for frontier metric "speed".}

\item{\code{activeness}}{List of rules for frontier metric "activeness".}

\item{\code{left}}{List of rules for frontier metric "woodland left".}
}}

