% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{FrontierMetric_summary-class}
\alias{FrontierMetric_summary-class}
\title{Class 'FrontierMetric_summary'}
\description{
Objects of class 'FrontierMetric_summary' are generated with \code{\link[=fmetrics_summary]{fmetrics_summary()}},
and hold summary statistics of frontier metrics, as well as the area occupied by
each frontier metric class.
}
\section{Slots}{

\describe{
\item{\code{summary_stats}}{Data frame with summary statistics.}

\item{\code{classes_areas}}{Data frame with the areas of frontier metric's classes.}

\item{\code{hists}}{'ggplot' object with frontier metric's histograms.}
}}

