% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gfw.R
\name{get_gfw}
\alias{get_gfw}
\title{Downloads and processes Global Forest Watch tiles}
\usage{
get_gfw(
  study_area,
  dir = ".",
  mask = TRUE,
  filenames = c("tree_cover", "loss_year"),
  overwrite = TRUE,
  silent = FALSE,
  ...
)
}
\arguments{
\item{study_area}{Either: (1) a 'SpatExtent' object defining the study area extent
in latitude/longitude coordinates; (2) a 'SpatVector' object with the polygon of the
study area; or (3) a file path to a spatial object.}

\item{dir}{Path to the directory where the processed raster layers will be saved. Default is current
directory.}

\item{mask}{Logical. If \code{TRUE} (default) downloaded tiles will be masked according to the
boundaries of the provided extent or polygon in \code{study_area}. Otherwise, the whole tiles will be kept.}

\item{filenames}{Character vector of length 2, indicating the name for the outputs of tree cover
and loss (raster layers). Default is \code{c("tree_cover", "loss_year")}.}

\item{overwrite}{Logical. If \code{TRUE}, files will be overwritten.}

\item{silent}{Logical. If \code{TRUE}, suppresses messages. Default is \code{FALSE}.}

\item{...}{Other arguments passed to \code{curl::multi_download}.}
}
\value{
Processed raster layers from GFW databases, one for tree canopu cover
and another for forest cover loss per year, based on the provided study area.
}
\description{
Based on the extent or polygon of a provided study area, downloads overlapping tiles from Global Forest Watch database
2024-v1.12 (Hansen et al. 2013), including tree canopy cover in year 2000 and forest cover loss per year until 2024,
and exports processed raster layers.
}
\details{
The function uses \code{curl::multi_download} to download multiple files (tiles) concurrently.
The downloaded tiles include:
(1) a raster layer of tree canopy cover in year 2000;
(2) a raster layer of forest cover loss for subsequent years up to 2024. The loss layer is
encoded as 0 (no loss) or as integer values (1, 2, ..., 24) corresponding to loss detected primarily in years 2001, 2001, ..., 2024.

Tiles of the same variable (tree canopy cover or forest loss) are merged and cropped to the
study area extent specified in \code{study_area}. If \code{mask = TRUE}, the tiles will be masked to
the provided extent or boundaries of the polygon.
}
\examples{
\dontrun{
# Creates working directory
dir.create("copo_tiles")

# Downloads study_area of study area
curl::curl_download(frontiermetrics_data[1], "copo.gpkg")

# Loads study_area of study area
copo <- terra::vect("copo.gpkg")

# Downloads GFW tiles and exports cropped and masked raster layers
get_gfw(study_area = copo, dir = "copo_tiles")
}
}
\references{
Hansen, M. C., Potapov, P. V., Moore, R., Hancher, M., Turubanova, S. A., Tyukavina, A.,
... & Townshend, J. R. (2013). High-resolution global maps of 21st-century forest cover change. science, 342(6160), 850-853.
}
