% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GFL_2024_v1.12_loss}
\alias{GFL_2024_v1.12_loss}
\title{Data frame containing the links to each tile of forest loss from Global Forest Watch (Hansen et al. 2013).
It is intended to be used for internal functions of the package.}
\format{
An object of class \code{data.frame} with 504 rows and 5 columns.
}
\usage{
GFL_2024_v1.12_loss
}
\description{
Data frame containing the links to each tile of forest loss from Global Forest Watch (Hansen et al. 2013).
It is intended to be used for internal functions of the package.
}
\references{
Hansen, M. C., Potapov, P. V., Moore, R., Hancher, M., Turubanova, S. A., Tyukavina, A.,
... & Townshend, J. R. (2013). High-resolution global maps of 21st-century forest cover change. science, 342(6160), 850-853.
}
\keyword{datasets}
