% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_fmetrics.R
\name{init_fmetrics}
\alias{init_fmetrics}
\title{Generates primary dataset to develop frontier metrics}
\usage{
init_fmetrics(
  gfw_cover,
  gfw_loss,
  min_treecover = 10,
  aggregation = c(10, 10),
  year_range = c(2000, 2023),
  min_cover = 5,
  min_rate = 0.5,
  window = 5,
  ncores = 1
)
}
\arguments{
\item{gfw_cover}{An object of class 'SpatRaster' or 'RasterLayer', or a path
(string) to a raster layer of tree cover from the GFW database.}

\item{gfw_loss}{An object of class 'SpatRaster' or 'RasterLayer', or a path
(string) to a raster layer of tree cover from the GFW database.}

\item{min_treecover}{A number (percentage) depicting the minimum percentage of tree
cover to be considered as "woodland". Default is 10\%. See Details.}

\item{aggregation}{A numeric vector of length two, depicting the first and second
magnitude of aggregation. See Details.}

\item{year_range}{A numeric vector of two elements depicting the first and last year of the analyzed time-frame.
Default is c(2000, 2023).}

\item{min_cover}{A number (percentage) depicting the minimum percentage of woodland
cover for a cell to be considered a frontier. Default is 5\%. See Details.}

\item{min_rate}{A number (percentage) depicting the minimum average annual woodland loss rate
in a five-year period for a cell to be considered a frontier. Default is 0.5\%. See Details.}

\item{window}{A whole number depicting the number of years to consider in the calculation of activeness.
Default is 5. See Details.}

\item{ncores}{Numbers of cores to parallelize processes.}
}
\value{
An object of class 'GFW_dataset', containing the primary dataset to
calculate frontier metrics with \code{\link[=fmetrics]{fmetrics()}}.
}
\description{
Based on inputs of tree cover and tree loss from Global Forest Watch database, generates
a structured dataset that can be used as main input to calculate frontier
metrics with function \code{\link[=fmetrics]{fmetrics()}}.
}
\details{
Global Forest Watch databases (Hansen et al. 2013) provide the percentage of tree cover in each
cell. One must define a minimum percentage of tree cover to a cell be considered
as "woodland". Usually, 10\% is used for subtropical dry forests, and 5\% for
tropical forests.

Aggregation is done in two steps. First, cells are aggregated using the first
value of the aggregation argument, which represents the number of cells to merge
in each direction (horizontal and vertical). In this step, the values of woodland cover of all
included cells are summed, and the amount of woodland loss is calculated for each new aggregated cell.
Second, the newly aggregated cells are further
grouped using the second value of the aggregation argument. At this stage,
each cell from the initial aggregation is merged into larger cells,
which will serve as the basis for calculating frontier metrics. Since GFW raster layers have a resolution of approximately 30m (at the equator),
setting \code{aggregation = c(10, 10)} (default setting) will: (1) aggregate 10 cells both horizontally
and vertically, producing a raster layer with a resolution of
~300m, and (2) further aggregate these newly aggregated cells by a factor of
10, creating larger cells with a resolution of ~3000m. For instance, setting
\code{aggregation = c(5, 10)} will first agregate to ~150m cells (30m * 5), then
aggregate these new cells into ~1500m cells (150m * 10).

Frontiers are defined as cells that meet two criteria: (1) they have a minimum
percentage of woodland cover (min_cover) and (2) they have a minimum average
annual woodland loss rate in a given temporal window (min_rate). By default, these
values take 5\% and 0.5\%, respectively (Buchadas et al. 2022). Those cells
that do not meet these criteria are excluded from the analysis. Also by default,
the temporal window is set to five years. However, this window could be set smaller
(e.g. 2 or 3 years) or bigger (e.g. 7 or 9 years).
}
\examples{
\dontrun{
# Loads raster layers generated with function get_gfw()
gfw_cover <- terra::rast(system.file("extdata", "tree_cover.tif", package = "frontiermetrics"))
gfw_loss <- terra::rast(system.file("extdata", "loss_year.tif", package = "frontiermetrics"))

# Generates primary dataset
copo_dataset <- init_fmetrics(gfw_cover, gfw_loss, year_range = c(2000, 2023), ncores = 2)
}
}
\references{
Buchadas, A., Baumann, M., Meyfroidt, P., & Kuemmerle, T. (2022). Uncovering major types of deforestation frontiers across the world's tropical dry woodlands. #' Nature Sustainability, 5(7), 619-627.

Hansen, M. C., Potapov, P. V., Moore, R., Hancher, M., Turubanova, S. A., Tyukavina, A.,
... & Townshend, J. R. (2013). High-resolution global maps of 21st-century forest cover change. science, 342(6160), 850-853.
}
