% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,FrontierMetric_breaks-method}
\alias{show,FrontierMetric_breaks-method}
\title{Shows 'FrontierMetric_breaks' object}
\usage{
\S4method{show}{FrontierMetric_breaks}(object)
}
\arguments{
\item{object}{Prints relevant information about a 'FrontierMetric_breaks' object.}
}
\value{
No return value.
}
\description{
Shows 'FrontierMetric_breaks' object
}
