% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,FrontierMetric_classes-method}
\alias{show,FrontierMetric_classes-method}
\title{Show 'FrontierMetric_classes' object}
\usage{
\S4method{show}{FrontierMetric_classes}(object)
}
\arguments{
\item{object}{Prints relevant information about a 'FrontierMetric_classes' object.}
}
\description{
Show 'FrontierMetric_classes' object
}
