% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,GFW_dataset-method}
\alias{show,GFW_dataset-method}
\title{Show 'GFW_dataset' object}
\usage{
\S4method{show}{GFW_dataset}(object)
}
\arguments{
\item{object}{Prints relevant information about a 'FrontierMetric' object.}
}
\description{
Show 'GFW_dataset' object
}
