% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks_rules.R
\name{breaks_rules}
\alias{breaks_rules}
\title{Sets the rules to categorize frontier metrics}
\usage{
breaks_rules(
  baseline = list("jenks", c("Low", "Medium", "High")),
  baseline_frag = list("jenks", c("Low", "Medium", "High")),
  loss = list("jenks", c("Low", "Medium", "High")),
  loss_frag = list("jenks", c("Low", "Medium", "High")),
  speed = list("jenks", c("Slow", "Medium", "Fast")),
  left = list("jenks", c("Low", "Medium", "High"))
)
}
\arguments{
\item{baseline, baseline_frag, loss, loss_frag, speed, left}{A list defining the
rules to define discrete classes for each frontier metric. See Details.}
}
\value{
An object of class 'FronterMetric_classes' to be used in argument \code{breaks} within \code{\link[=fmetrics]{fmetrics()}}.
}
\description{
Defines rules to generate discrete categories for continuous frontier metrics.
}
\details{
For each metric, the user must provide a list. The first element of the list
defines the algorithm to be used for the categorization. The second element of  the list must be a character vector with the names
for each category that will be generated.

The following algoritms are available (first element of the list):

(1) "jenks" (default): Jenks natural breaks classification (uses BAMMtools::getJenksBreaks).
The number (n) of intervals is given by the number of categories provided
in the second element of the list.

(2) "equal": splits the numeric range between minimum and maximum metric values into n intervals of equal
width. The number (n) of intervals is given by the number of categories provided
in the second element of the list.

(3) "quantile": splits observations by quantiles so that each bin contains (roughly)
the same number of observations (frontier cells). Quantiles are calculated given
the number of categories provided in the second element of the list. For instance,
if 4 categories are provided, the following intervals be used as cuts: 0-0.25, 0.25-0.5, 0.5-0.75, and 0.75-1.

(4) Alternatively, a numeric vector with two or more unique cut points can be
provided. For example: \code{c(5, 10, 55, Inf)}, and the categories (second element
of the list) could be \code{c("Low", "Medium", "High")}, where "Low": 5-10, "Medium": 10-55, and
"High": 55-Inf.

When assigned to an object, the new definition of classes can be passed to \code{\link[=fmetrics]{fmetrics()}},
inside the argument \code{breaks}, to generate frontier metrics's classes with a new definition.
}
\examples{
# Explore default rules
breaks_rules()

# Define new parameters for baseline
#   Low: between 5 and 20
#   Medium: between 20 and 70
#   High: Higher than 70
breaks_1 <- breaks_rules(baseline = list(c(5, 20, 70, Inf),
                                         c("Low", "Medium", "High")))

# Shows new defined rules
breaks_1
}
