% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmetrics.R
\name{fmetrics}
\alias{fmetrics}
\title{Calculates frontier metrics}
\usage{
fmetrics(
  x,
  metrics = "all",
  classes = init_classes(x),
  ncores = 1,
  silent = FALSE
)
}
\arguments{
\item{x}{An object of class 'GFW_dataset' generated with \code{\link[=init_fmetrics]{init_fmetrics()}}.}

\item{metrics}{Frontier metrics to be calculated. Options are: "baseline",
"loss", "fragmentation", "speed", "activeness" and/or "left" for individual
frontier metrics; "severity", "spatio_temporal" and/or "development" for
frontier typologies; "all" to calculate all frontier metrics. Default is "all".
See Details.}

\item{classes}{An object of class 'FrontierMetric_classes' containing the rules
to define discrete classes for individual frontier metrics. See Details.}

\item{ncores}{Numbers of cores to parallelize processes. Default is 1. See Details.}

\item{silent}{Logical. If TRUE, suppresses messages. Default is FALSE.}
}
\value{
An object of class 'FrontierMetric', containing a dataset with the calculated frontier metrics
for each individual cell. This object can be passed to \code{\link[=fmetrics_summary]{fmetrics_summary()}}, \code{\link[=fmetrics_plot]{fmetrics_plot()}}, and \code{\link[=fmetrics_rast]{fmetrics_rast()}}.
}
\description{
Given a structured dataset of class 'GFW_dataset', generated with \code{\link[=init_fmetrics]{init_fmetrics()}},
calculates frontier metrics following the methodology of Buchadas et al. (2022).
}
\details{
The \code{metrics} argument specifies which frontier metrics will be calculated. These metrics are computed for each cell and are defined as follows:
\itemize{
\item baseline: The percentage of woodland cover in each cell in the first year of the time-frame.
\item loss: The percentage of total woodland loss over the studied time-frame, relative to the baseline woodland cover.
\item fragmentation: The maximum value of edge density (m/ha) per year within the studied time-frame.
\item speed: The maximum rate of woodland loss (km²/year) during the studied time-frame.
\item activeness: Categorized as "active", "old", or "emerging", depending on when the frontier was active through
five consecutive year-periods during the period of analysis.
\item left: The percentage of woodland left after the period of woodland loss analysed.
}

For detailed explanations of how frontier metrics are calculated, refer to Buchadas et al. (2022).
To improve flexibility, frontier metrics "baseline", "loss", and "left" are expressed as a percentages, rather than in km².

The \code{metrics} argument also allows specifying groups of metrics or "frontier typologies"
(Buchadas et al. 2022). Typologies include:
"severity" (comprising the "baseline" and "loss" metrics),
"spatio_temporal" (comprising the "speed" and "fragmentation" metrics), and
"development" (comprising the "activeness" and "left" metrics).

The rules to define discrete categories for frontier metrics can
be defined in argument \code{classes}. This accepts an object of class 'FrontierMetric_classes',
generated with \code{\link[=init_classes]{init_classes()}}. Run \code{init_classes(x)} with \code{x} equal to an object
of class 'GFW_dataset' to explore these rules and see ?init_classes to
explore how to re-define these categories.

By default, in the scenario of default definition of frontiers, rules to
categorize frontier metrics are the following:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Baseline woodland [\%]:
  Low (5,10]
  Medium (10,55]
  High (55,Inf]

Woodland loss [\%]:
  Low (-Inf,20]
  Medium (20,50]
  High (50,Inf]

Fragmentation [m/ha]:
  Low (-Inf,10]
  Medium (10,20]
  High (20,Inf]

Speed [km²/year]:
  Slow (-Inf,0.05]
  Medium (0.05,0.25]
  Fast (0.25,Inf]

Activeness:
  Old       ... ,[2012,2016], [2013,2017], [2014,2018]
  Active    [2015,2019], [2016,2020] ... [2018,2022]
  Emerging  [2019,2023]

Woodland left [\%]:
  Low (-Inf,33]
  Medium (33,66]
  High (66,Inf]
}\if{html}{\out{</div>}}

Parallelization is recommended for very large datasets (ncores > 1).
The calculation of the fragmentation metric is the most computationally intensive.
In general, it is advisable to first run the function with \code{ncores = 1} and see if it is feasible.
If needed, increase the number of cores. The calculation of this particular metric with
parallelization requires the \code{snowfall} package.
}
\examples{
\dontrun{
# Calculates all frontier metrics
copo_metrics <- fmetrics(copo_dataset, metrics = "all", ncores = 2)

# Calculates development frontiers
copo_metrics2 <- fmetrics(copo_dataset, metrics = "development", ncores = 2)
copo_metrics2

# Calculates a single frontier
copo_metrics3 <- fmetrics(copo_dataset, metrics = "speed", ncores = 2)
copo_metrics3
}
}
\references{
Buchadas, A., Baumann, M., Meyfroidt, P., & Kuemmerle, T. (2022). Uncovering major types of
deforestation frontiers across the world’s tropical dry woodlands. Nature Sustainability, 5(7), 619-627.
}
\seealso{
\code{\link[=init_fmetrics]{init_fmetrics()}}, \code{\link[=fmetrics_summary]{fmetrics_summary()}}, \code{\link[=fmetrics_plot]{fmetrics_plot()}}, and \code{\link[=fmetrics_rast]{fmetrics_rast()}}
}
