% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmetrics.R
\name{fmetrics}
\alias{fmetrics}
\title{Calculates frontier metrics}
\usage{
fmetrics(
  x,
  metrics = "all",
  params = list(activeness_levels = NULL, onset_min_years = 3),
  breaks = breaks_rules(),
  dir = NULL,
  gdal = NULL,
  overwrite = TRUE,
  export_archetypes = FALSE,
  ncores = 1,
  silent = FALSE
)
}
\arguments{
\item{x}{An object of class 'init_FrontierMetric' generated with \code{\link[=init_fmetrics]{init_fmetrics()}}.}

\item{metrics}{Frontier metrics to be calculated. Options are: "baseline",
"baseline_frag", "activeness", "speed", "loss", "left", "loss_frag" and "onset";
"all" (default) to calculate all available frontier metrics. User-defined metrics
can also be calculated. See Details.}

\item{params}{List of parameters related to the levels of activeness of a
frontier and the minimum years to consider for the 'onset' metric. See Details.}

\item{breaks}{An object of class 'FrontierMetric_breaks' generated with \code{\link[=breaks_rules]{breaks_rules()}}, containing the rules
to define discrete classes for individual frontier metrics. See Details.}

\item{dir}{A path to a directory to export raster layers of frontier metrics. If \code{NULL},
(default) raster layers will not be exported. See Details.}

\item{gdal}{GDAL driver specific datasource creation options. See ?terra::writeRaster for details.}

\item{overwrite}{Logical. If \code{TRUE} (default), raster layers files will be overwritten.}

\item{export_archetypes}{Logical. If \code{TRUE} and a path is provided in \code{dir},
raster layer of archetypes will also be exported. Default is \code{FALSE}.}

\item{ncores}{Numbers of cores to parallelize processes. Default is 1. See Details.}

\item{silent}{Logical. If \code{TRUE}, suppresses messages. Default is \code{FALSE}.}
}
\value{
An object of class 'FrontierMetric', containing a dataset with the calculated frontier metrics
for each individual cell. This object can be passed to \code{\link[=fmetrics_summary]{fmetrics_summary()}},
\code{\link[=fmetrics_plot]{fmetrics_plot()}}, and \code{\link[=fmetrics_rast]{fmetrics_rast()}}.
}
\description{
Given a structured object of class 'init_FrontierMetric' generated with \code{\link[=init_fmetrics]{init_fmetrics()}},
calculates a set of deforestation frontier metrics.
}
\details{
The \code{metrics} argument specifies which frontier metrics will be calculated. These metrics are computed for each cell and are defined as follows:
\itemize{
\item baseline: The percentage of forest cover in each cell in the first year of the time-frame.
\item baseline_frag: The value of edge density (m/ha) in the first year of the time-frame.
\item loss: The percentage of total forest loss over the studied time-frame, relative to the baseline forest cover.
\item loss_frag: The maximum value of edge density (m/ha) per year of the forest loss spatial pattern, within the studied time-frame.
\item speed: The maximum rate of forest loss (km²/year) during the studied time-frame.
\item activeness: The level of activeness, as discrete categories, according to when the frontier was active
during the period of analysis.
\item left: The percentage of forest cover left after at the end of the studied time-series, relative to the
area of the cell.
\item onset: The year of onset of the deforestation frontier, calculated as the
first year with at least 3 years (by default) of consecutive forest loss.
}

Key parameters of "activeness" and "onset" can be changed within the argument \code{params}.
By default, \code{params = list(activeness_levels = NULL, onset_min_years = 3)}.  The level
of frontier activeness will depend on the time-windows where the frontier
was active along the time-frame. Time-windows can be visualized, before calculating frontiers,
by inspecting the slot \verb{@temporal_windows} of the object of class 'init_FrontierMetric' generated
with \code{init_fmetrics()} and provided in argument \code{x} of \code{fmetrics()}. For instance,
if we consider the time-frame 2000-2024 and a window of 5 years, the
temporal windows for this time-frame will be:\tabular{lll}{
   window     \tab first_year     \tab last_year     \cr
   1 \tab 2001 \tab 2005 \cr
   2 \tab 2002 \tab 2006 \cr
   3 \tab 2003 \tab 2007 \cr
   4 \tab 2004 \tab 2008 \cr
   5 \tab 2005 \tab 2009 \cr
   6 \tab 2006 \tab 2010 \cr
   7 \tab 2007 \tab 2011 \cr
   8 \tab 2008 \tab 2012 \cr
   9 \tab 2009 \tab 2013 \cr
   10 \tab 2010 \tab 2014 \cr
   11 \tab 2011 \tab 2015 \cr
   12 \tab 2012 \tab 2016 \cr
   13 \tab 2013 \tab 2017 \cr
   14 \tab 2014 \tab 2018 \cr
   15 \tab 2015 \tab 2019 \cr
   16 \tab 2016 \tab 2020 \cr
   17 \tab 2017 \tab 2021 \cr
   18 \tab 2018 \tab 2022 \cr
   19 \tab 2019 \tab 2023 \cr
   20 \tab 2020 \tab 2024 \cr
}


If \code{activeness_levels = NULL}, frontier activeness will be classified as "emerging" (active
during temporal window 20), "active" (active during temporal windows 16 to 19) or
"old" (active during temporal windows 1 to 15). These categories can be changed. For instance,
if \code{activeness_levels = list(emerging = 19:20, old = 15:18, very_old = 1:14)}, three levels of
activeness will be considered, according to the provided temporal windows for each one.
All temporal windows must be considered in the defined activeness levels.

If \code{onset_min_years = 3} (default), onset will refer to the first year of the time series
that exhibited forest loss in a number of 3 consecutive years. The user can change this
parameter to a lower or higher minimum number of years.

\emph{
Note that these metrics were initially programmed to work with Global Forest Watch
datasets, and excluding the process of forest gain. This could be relevant for the
calculation of some of these metrics, if a custom time-series of forest cover is
provided, and the process of forest gain is considered within this time-series. In
addition, these metrics where initially thought for the process of forest loss in
tropical and subtropiccal regions. Any other use for a different cover or region
should be applied with caution, and these metrics might need particular adjustments.
}

For advanced R users, user-defined metrics can also be calculated. The user must
create an R function (named "ud_" + "metric_name"; e.g. "ud_metric1") that would receive a unique argument \code{x}, an object of class
\code{init_FrontierMetric} generated with \code{init_fmetrics()}, work with the structured
dataset within this object as desired (available in slot \verb{@data}), and return a data.frame
with the values of the metric for each individual cell. See the vignette for details and
an example.

In argument \code{breaks}, the user can define the rules to generate discrete
classes of those continuous metrics, by providing an object of
class 'FrontierMetric_breaks' generated with \code{\link[=breaks_rules]{breaks_rules()}}. Run \code{breaks_rules()} to
explore these rules. By default, continuous metrics
are divided in discrete classes by using the "Jenks natural
breaks classification", using the function \code{BAMMtools::getJenksBreaks}.
See ?breaks_rules to explore how to re-define categories.

Based on the calculated frontier metrics, the function also calculates
frontier archetypes. These can be defined as the combination
of different frontier metric classes that might represent a higher order of
frontier classification. Archetypes can be accessed through \verb{@archetypes} in
the outputted object.

If a path is provided in \code{dir} (default is \code{NULL}), raster layers of frontier metrics will be
exported as .tif files. Use \code{dir = ""} to export files to current directory.
GDAL options can be defined in argument \code{gdal}. See ?terra::writeRaster for details.

Parallelization is recommended for very large datasets (ncores > 1).
The calculation of the the metric "loss_frag" is the most computationally intensive.
In general, it is advisable to first run the function with \code{ncores = 1} and see if it is feasible.
If needed, increase the number of cores. The calculation of this particular metric with
parallelization requires the \code{snowfall} package to be installed.
}
\examples{
# Downloads example of object of class 'init_FrontierMetric' generated with init_fmetrics()
curl::curl_download(frontiermetrics_data[4], file.path(tempdir(), "copo_dataset.RDS"))

# Loads object to R environment
copo_dataset <- readRDS(file.path(tempdir(), "copo_dataset.RDS"))

# Calculates specific metrics;
#   Use "all" for all available metrics;
#   Use dir = "path_to_directory" to also export as raster layers
copo_metrics2 <- fmetrics(copo_dataset, metrics = c("baseline", "speed", "left", "onset"))

# Inspect generated object
copo_metrics2
}
\references{
Buchadas, A., Baumann, M., Meyfroidt, P., & Kuemmerle, T. (2022). Uncovering major types of
deforestation frontiers across the world’s tropical dry woodlands. Nature Sustainability, 5(7), 619-627.
}
\seealso{
\code{\link[=init_fmetrics]{init_fmetrics()}}, \code{\link[=fmetrics_summary]{fmetrics_summary()}}, \code{\link[=fmetrics_plot]{fmetrics_plot()}}, and \code{\link[=fmetrics_rast]{fmetrics_rast()}}
}
