% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmetrics_plot.R
\name{fmetrics_plot}
\alias{fmetrics_plot}
\title{Plots frontier metrics}
\usage{
fmetrics_plot(
  x,
  metrics = "all",
  type = "both",
  ncol = 4,
  palette = "viridis",
  direction = c(1, 1),
  background = c("gray90", "gray64")
)
}
\arguments{
\item{x}{Object of class 'FrontierMetric' generated with \code{\link[=fmetrics]{fmetrics()}}}

\item{metrics}{Frontier metrics to be plotted. Options are: "baseline", "loss",
"fragmentation", "speed", "activeness" and/or "left" for individual frontier
metrics; "severity", "spatio_temporal" and/or "development" for frontier typologies;
"all" to calculate all frontier metrics available within \code{x}. Default is "all".}

\item{type}{One of the following: "both", to plot absolute values and discrete
classes of frontier metrics; "values", to plot only absolute values; or "classes"
to plot only discrete classes. Default is "both".}

\item{ncol}{Number of columns to arrange plots. Default is 4.}

\item{palette}{A vector of one or two strings with the names of the palettes to be used
in the plots. If two strings are provided, the first element is for the absolute values of frontier metrics and the second for the
discrete classes. Default is \code{"viridis"}. If only one palette is provided,
the same palette will be used for both plots. See Details.}

\item{direction}{A vector of two integers indicating the direction of the palette
for the absolute values of frontier metrics and for discrete classes, respectively.
Integers must be 1 or -1. If only one number is provided, the same direction will
be used for both palettes. Default is c(1, 1). See Details.}

\item{background}{A vector of two strings with the names of the colors to be used
for the background of the plot. The first depicts the color for the region outside
the range of the study area, while the second depicts the color for those cells
within the study area that did not qualify as frontiers. Default is c("gray90", "gray64").
See Details.}
}
\description{
Generates plots of frontier metrics based on an object of class 'FrontierMetric'
generated with \code{\link[=fmetrics]{fmetrics()}}
}
\details{
Available palettes are those available for the function \code{scale_fill_viridis()}:
"magma", "inferno", "plasma", "viridis", "cividis", "rocket", "mako", or "turbo".
The \code{direction} argument is used to reverse the palette if necessary. To do this
the user must specify -1. The user can play with the background colors for
better visualization.
}
\examples{
\dontrun{
fmetrics_plot(copo_metrics)
fmetrics_plot(copo_metrics, metrics = "speed", type = "values", palette = "magma")
}
}
