% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmetrics_rast.R
\name{fmetrics_rast}
\alias{fmetrics_rast}
\title{Exports raster layers of frontier metrics}
\usage{
fmetrics_rast(x, metrics = "all", dir = NULL, gdal = NULL, overwrite = TRUE)
}
\arguments{
\item{x}{An object of class 'FrontierMetric' generated with \code{\link[=fmetrics]{fmetrics()}}}

\item{metrics}{Frontier metrics to be exported as raster layers. Options are: "baseline", "loss",
"fragmentation", "speed", "activeness", and/or "left" for individual frontier
metrics; "severity", "spatio_temporal", and/or "development" for frontier typologies; "archetypes"
for frontier archetypes; "all" for all frontier metrics and archetypes available within \code{x}.
Default is "all".}

\item{dir}{Path to the directory where to export raster layers.}

\item{gdal}{GDAL driver specific datasource creation options. See the GDAL documentation.}

\item{overwrite}{Logical. If \code{TRUE}, files will be overwritten.}
}
\value{
Invisible. Exports raster layers as .tif files. Absolute values, discrete classes
of each metric, and frontier archetypes are exported as separate layers within each .tif file.
A text file named README.txt with details on the discrete classes, and another text file
named ARCHETYPES.txt with a table with definitions of the archetypes, are also created.
}
\description{
Based on an object of class 'FrontierMetric', exports raster layers of
frontier metrics, both the absolute values and the
discrete classes, as well as frontier archetypes.
}
\details{
Raster layers are exported as .tif files, with the following names:
"baseline_forest.tif", "forest_loss.tif,"speed.tif", "fragmentation.tif",
"activeness.tif", and "forest_left.tif". Raster layer of archetypes is exported
as "archetypes.tif". In addition, a raster layer named "is_frontier.tif" is
also exported, which shows cells that are frontiers (value = 1) and those that
are not (value = 0).
}
\examples{
\dontrun{
fmetrics_rast(copo_metrics)
}
}
