% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmetrics_rast.R
\name{fmetrics_rast}
\alias{fmetrics_rast}
\title{Exports raster layers of frontier metrics}
\usage{
fmetrics_rast(
  x,
  metrics = "all",
  dir = "",
  gdal = NULL,
  overwrite = TRUE,
  silent = FALSE
)
}
\arguments{
\item{x}{An object of class 'FrontierMetric' generated with \code{\link[=fmetrics]{fmetrics()}}.}

\item{metrics}{Frontier metrics to be exported. Options include those calculated
frontier metrics (including user-defined metrics), "archetypes"
for frontier archetypes, or "all" for all frontier metrics and archetypes available within \code{x}.
Default is "all".}

\item{dir}{Path to the directory where raster layers will be exported. By default,
the current directory.}

\item{gdal}{GDAL driver specific datasource creation options. See ?terra::writeRaster for details.}

\item{overwrite}{Logical. If \code{TRUE}, files will be overwritten.}

\item{silent}{Logical. If \code{TRUE}, suppresses messages. Default is \code{FALSE}.}
}
\value{
Invisible. Exports raster layers of metrics and archetypes as .tif files.
The function also creates a text file named README.txt with details on the discrete classes, and a text file
named archetypes.txt with a table containing the definitions of the archetypes.
}
\description{
Based on an object of class 'FrontierMetric', exports raster layers of
frontier metrics, both the continuous values and the
discrete classes, as well as frontier archetypes.
}
\details{
Raster layers are exported as .tif files named after each exported metric, including
continuous and discrete classes within each raster layer. Raster layer of archetypes is exported
as "archetypes.tif". In addition, a raster layer named "is_frontier.tif" is
also exported, which depict cells classified as frontiers (value = 1) and those that
were not (value = 0).

This function is also used by \code{\link[=fmetrics]{fmetrics()}} to export raster layers of frontier metrics and
archetypes, if a path to a directory is defined when running this function. If not,
\code{\link[=fmetrics_rast]{fmetrics_rast()}} can be directly used for this purpose.
}
\examples{
\dontrun{
# Exports all available metrics as raster layers, including archetypes
fmetrics_rast(copo_metrics)

# Exports a single metric as a raster layer
fmetrics_rast(copo_metrics, metrics = "left")

# Exports only archetypes
fmetrics_rast(copo_metrics, metrics = "archetypes")
}
}
