% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmetrics_summary.R
\name{fmetrics_summary}
\alias{fmetrics_summary}
\title{Calculates summary statistics of frontier metrics}
\usage{
fmetrics_summary(x, metrics = "all")
}
\arguments{
\item{x}{Object of class 'FrontierMetric' generated with \code{\link[=fmetrics]{fmetrics()}}.}

\item{metrics}{Frontier metrics to be plotted. Options include those calculated
frontier metrics (including user-defined metrics) or "all" to calculate all frontier
metrics available within \code{x}. Default is "all".}
}
\value{
A list with three elements:
(1) a data frame with basic summary statistics;
(2) a data frame with the total area of each class within each frontier metric; and
(3) a ggplot object with histograms of frontier metrics.
}
\description{
Calculates basic statistics of continuous values of frontier metrics, plots
histograms, and calculates the total area of each class within each frontier metric.
}
\details{
User-defined metrics can not be passed to this function.
}
\examples{
summ <- fmetrics_summary(copo_metrics)
summ

# Access to summary statistics
summ@summary_stats

# Access to classes areas
summ@classes_areas

# Access and plotting of histograms
summ@hists
}
