% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmetrics_summary.R
\name{fmetrics_summary}
\alias{fmetrics_summary}
\title{Calculates summary statistics of frontier metrics}
\usage{
fmetrics_summary(x, metrics = "all")
}
\arguments{
\item{x}{Object of class 'FrontierMetric' generated with \code{\link[=fmetrics]{fmetrics()}}.}

\item{metrics}{Frontier metrics to consider. Options are: "baseline", "loss",
"fragmentation", "speed", "activeness" and/or "left" for individual frontier
metrics; "severity", "spatio_temporal" and/or "development" for frontier typologies;
"all" to calculate all frontier metrics available within \code{x}. Default is "all".}
}
\value{
A list with three elements: (1) a data frame with basic summary statistics;
(2) a data frame with the total area of each class within each frontier metric; and
(3) a ggplot object with histograms of frontier metrics.
}
\description{
Calculates basic statistics of absolute values of frontier metrics and plots
histograms. Also, it calculates the total area of each class within each frontier metric.
}
\examples{
\dontrun{
summ <- fmetrics_summary(copo_metrics)
summ

# Access to summary statistics
summ@summary_stats

# Access to classes areas
summ@classes_areas

# Access and plotting of histograms
summ@hists
}
}
