% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frontiermetrics_data}
\alias{frontiermetrics_data}
\title{URLs to download data for example use.}
\format{
An object of class \code{character} of length 5.
}
\usage{
frontiermetrics_data
}
\description{
URLs to download data for example use.
}
\keyword{datasets}
