% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gfw.R
\name{get_gfw}
\alias{get_gfw}
\title{Downloads and processes Global Forest Watch tiles}
\usage{
get_gfw(polygon, mask = TRUE, dir = "", overwrite = FALSE, timeout = 600)
}
\arguments{
\item{polygon}{An object of class 'SpatVector' or a path to a vector file, representing the polygon
of the study area.}

\item{mask}{Logical. If TRUE (default) downloaded tiles will be cropped and masked according to the
boundaries of the provided polygon. Otherwise, the whole tiles will be kept.}

\item{dir}{Path to the directory where the processed raster layers will be saved. Default is current
directory.}

\item{overwrite}{Logical. If \code{TRUE}, files will be overwritten.}

\item{timeout}{Maximum time (in seconds) invested in downloading the tiles. Default is 600.}
}
\value{
Exports cropped and masked raster layers from GFW databases, one for tree cover
and another for woodland loss, based on the polygon provided in argument \code{polygon}.
}
\description{
Downloads all overlapping tiles from Global Forest Watch database (Hansen et al. 2013), based on the
extent of a user-provided polygon, and exports raster layers of tree cover and woodland loss.
}
\details{
Downloaded tiles are classified as follows: (1) raster layer depicting the tree cover
in the year 2000, and (2) raster layer depicting the woodland loss in the following years
until 2023. Tiles of the same type (tree cover or woodland loss)
are merged, then cropped and masked based on the on the polygon provided in
argument \code{polygon}.
}
\examples{
\dontrun{
# Creates working directory
dir.create("copo_tiles/")

# Downloads polygon of study area
download.file(frontiermetrics_data[1], dir = "copo.gpkg")

# Loads polygon of study area
copo <- terra::vect("copo.gpkg")

# Downloads GFW tiles and exports cropped and masked raster layers
get_gfw(polygon = copo, dir = "copo_tiles/")
}
}
\references{
Hansen, M. C., Potapov, P. V., Moore, R., Hancher, M., Turubanova, S. A., Tyukavina, A.,
... & Townshend, J. R. (2013). High-resolution global maps of 21st-century forest cover change. science, 342(6160), 850-853.
}
