% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gfw.R
\name{get_gfw}
\alias{get_gfw}
\title{Downloads and processes Global Forest Watch tiles}
\usage{
get_gfw(
  study_area,
  dir = "",
  mask = TRUE,
  overwrite = FALSE,
  timeout = 600,
  silent = FALSE
)
}
\arguments{
\item{study_area}{Either: (1) a 'SpatExtent' object defining the study area extent
in latitude/longitude coordinates; (2) a 'SpatVector' object with the polygon of the
study area; or (3) a file path to a spatial object.}

\item{dir}{Path to the directory where the processed raster layers will be saved. Default is current
directory.}

\item{mask}{Logical. If TRUE (default) downloaded tiles will be masked according to the
boundaries of the provided extent or polygon in \code{study_area}. Otherwise, the whole tiles will be kept.}

\item{overwrite}{Logical. If \code{TRUE}, files will be overwritten.}

\item{timeout}{Maximum time (in seconds) invested in downloading the tiles. Default is 600.}

\item{silent}{Logical. If \code{TRUE}, suppresses messages. Default is \code{FALSE}.}
}
\value{
Exports cropped and/or masked raster layers from GFW databases, one for tree cover
and another for forest loss, based on the provided study area.
}
\description{
Downloads all overlapping tiles from Global Forest Watch database 2024-v1.12 (Hansen et al. 2013), based on the
extent of a provided study area, and exports raster layers of tree cover and forest loss.
}
\details{
The downloaded tiles include: (1) a raster layer of tree cover for the year 2000,
and (2) raster layer of forest loss for subsequent years up to 2024.
Tiles of the same variable (tree cover or forest loss) are merged and cropped to the
study area extent specified in \code{study_area}. If \code{mask = TRUE}, the tiles will be masked to
the provided extent or polygon.
}
\examples{
\dontrun{
# Creates working directory
dir.create("copo_tiles")

# Downloads study_area of study area
curl::curl_download(frontiermetrics_data[1], "copo.gpkg")

# Loads study_area of study area
copo <- terra::vect("copo.gpkg")

# Downloads GFW tiles and exports cropped and masked raster layers
get_gfw(study_area = copo, dir = "copo_tiles")
}
}
\references{
Hansen, M. C., Potapov, P. V., Moore, R., Hancher, M., Turubanova, S. A., Tyukavina, A.,
... & Townshend, J. R. (2013). High-resolution global maps of 21st-century forest cover change. science, 342(6160), 850-853.
}
