% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_fmetrics.R
\name{init_fmetrics}
\alias{init_fmetrics}
\title{Generates primary dataset to develop frontier metrics}
\usage{
init_fmetrics(
  raster,
  is_series = FALSE,
  min_treecover = 10,
  aggregation = c(10, 10),
  year_range = c(2000, 2023),
  min_cover = 5,
  min_rate = 0.5,
  window = 5,
  ncores = 1
)
}
\arguments{
\item{raster}{If \code{is_series = FALSE}, a list of two objects of class 'SpatRaster',
or two paths to two raster layers, representing tree cover and woodland loss from
Global Forest Watch databases. If \code{is_series = TRUE}, an object of class 'SpatRaster', or a path
to a raster layer, representing a cover time-series. See Details.}

\item{is_series}{Logical. If \code{FALSE} (default), it is expected that the raster
layers provided in \code{raster} were obtained from Global Forest Watch databases.
If \code{TRUE}, it is expected that a cover time-series is provided in \code{raster}. See Details.}

\item{min_treecover}{A number (percentage) depicting the minimum percentage of tree
cover to be considered as "woodland". Default is 10\%. Ignored if \code{is_series = TRUE}. See Details.}

\item{aggregation}{A numeric vector of length two, depicting the first and second
magnitude of aggregation. See Details.}

\item{year_range}{A numeric vector of two elements depicting the first and last year of the analyzed time-frame.
Default is c(2000, 2023). Ignored if \code{is_series = TRUE}. See Details.}

\item{min_cover}{A number (percentage) depicting the minimum percentage of woodland
cover for a cell to be considered a frontier. Default is 5\%. See Details.}

\item{min_rate}{A number (percentage) depicting the minimum average annual woodland loss rate
in a five-year period for a cell to be considered a frontier. Default is 0.5\%. See Details.}

\item{window}{A whole number depicting the number of years to consider in the calculation of activeness.
Default is 5. See Details.}

\item{ncores}{Numbers of cores to parallelize processes.}
}
\value{
An object of class 'init_FrontierMetric', containing the primary dataset to
calculate frontier metrics with \code{\link[=fmetrics]{fmetrics()}}.
}
\description{
Based on inputs of tree cover and woodland loss from Global Forest Watch database, or
on a custom cover time-series, generates a structured dataset that can be used as the main input to calculate frontier
metrics with \code{\link[=fmetrics]{fmetrics()}}.
}
\details{
Deforestation frontier metrics were initially developed to be calculated with Global Forest Watch databases
(Buchadas et al. 2022; Hansen et al. 2013). If data from Global Forest Watch is used, \code{is_series} must equal
\code{FALSE}, the value by default. Then, a list with two main raster layers from this source must be provided,
in the following order:

(1) a raster layer (class SpatRaster or a path to a raster layer) of initial woodland cover in year 2000.
The values for this raster layer typically ranges
between 0 and 100, depicting the percentage of tree cover of each individual cell
(at a resolution of ~30m).

(2) a raster layer (class SpatRaster or a path to a raster layer) of woodland loss between years 2001
and 2023. This is a single raster layer where 0 indicates no loss, while values of 1, 2, 3, etc.,
represent woodland loss occurring in years 1, 2, 3, and so on.

Both raster layers can be previously downloaded and processed with \code{\link[=get_gfw]{get_gfw()}}, which will
access the Global Forest Watch database. See \code{?get_gfw} for detailed
documentation. One must also define a minimum percentage of tree cover to a cell be considered
"woodland". Usually, 10\% is used for subtropical dry forests, and 5\% for
tropical forests.

Alternatively, a custom raster layer representing a time-series of
woodland cover can be provided. If so, \code{is_series} must equal \code{TRUE}, and
a raster of multiple layers must be provided in argument \code{raster}.
Within this object, each layer must be binary, representing the amount of woodland cover for
each consecutive year of the analyzed time series. A cell of value 1 will be
considered as covered by woodland, while 0 will be considered as a different
cover. No other values, besides 0, 1, and NA should be present in each layer.
In addition, when \code{is_series = TRUE}, argument \code{year_range} must
represent the first and last year of the time-series. The number of years
given by this range must match the number of layers of the cover time-series
provided in \code{raster}.

Aggregation is done in two steps. First, cells are aggregated using the first
value of the aggregation argument, which represents the number of cells to merge
in each direction (horizontal and vertical). In this step, the values of woodland cover of all
included cells are summed, and the amount of woodland loss is calculated for each new aggregated cell.
Second, the newly aggregated cells are further
grouped using the second value of the aggregation argument. At this stage,
each cell from the initial aggregation is merged into larger cells,
which will serve as the basis for calculating frontier metrics. Since GFW raster layers have a resolution of approximately 30m (at the equator),
setting \code{aggregation = c(10, 10)} (default setting) will: (1) aggregate 10 cells both horizontally
and vertically, producing a raster layer with a resolution of
~300m, and (2) further aggregate these newly aggregated cells by a factor of
10, creating larger cells with a resolution of ~3000m. For instance, setting
\code{aggregation = c(5, 10)} will first agregate to ~150m cells (30m * 5), then
aggregate these new cells into ~1500m cells (150m * 10).

Frontiers are defined as cells that meet two criteria: (1) they have a minimum
percentage of woodland cover (\code{min_cover}) and (2) they have a minimum average
annual woodland loss rate in a given temporal window (\code{min_rate}). By default, these
values take 5\% and 0.5\%, respectively (Buchadas et al. 2022), but they could
take other values. Those cells
that do not meet these criteria are excluded from the analysis. Also by default,
the temporal window is set to five years, but it could also be set to a different
number of years(e.g. 2, 3, 7 or 9 years).
}
\examples{
\dontrun{
# Downloads raster layers generated with function get_gfw()
# Tree cover
download.file(frontiermetrics_data[2], "tree_cover.tif")
# Woodland loss
download.file(frontiermetrics_data[3], "loss_year.tif")

# Loads raster layers generated with function get_gfw()
rast_cover <- terra::rast("tree_cover.tif")
rast_loss <- terra::rast("loss_yearr.tif")

# Generates primary dataset
copo_dataset <- init_fmetrics(raster = list(rast_cover, rast_loss), ncores = 2)
}
}
\references{
Buchadas, A., Baumann, M., Meyfroidt, P., & Kuemmerle, T. (2022). Uncovering major types of deforestation frontiers across the world's tropical dry
woodlands. Nature Sustainability, 5(7), 619-627.

Hansen, M. C., Potapov, P. V., Moore, R., Hancher, M., Turubanova, S. A., Tyukavina, A.,
... & Townshend, J. R. (2013). High-resolution global maps of 21st-century forest cover change. science, 342(6160), 850-853.
}
