% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{init_FrontierMetric-class}
\alias{init_FrontierMetric-class}
\title{Class 'init_FrontierMetric'}
\description{
Objects of class 'init_FrontierMetric' are generated with \code{\link[=init_fmetrics]{init_fmetrics()}}, and hold
the necessary objects to calculate frontier metrics with \code{\link[=fmetrics]{fmetrics()}}. The main
object is under \verb{@data}, which is a data frame containing -among other relevant data- the amount of forest
cover in the first year of the time series, and the amount of forest loss
in each year of the time series, for each individual cell of the study area.
}
\section{Slots}{

\describe{
\item{\code{data}}{Main dataset containing data about forest cover and forest loss.}

\item{\code{coords}}{Central coordinates of bigger cells.}

\item{\code{time_frame}}{Year range of the studied time series.}

\item{\code{initial_fc_col}}{Name of the column containing the data about initial forest cover.}

\item{\code{fl_cols}}{Name of the columns containing the data about annual forest loss.}

\item{\code{extent}}{Extent of the study area.}

\item{\code{grain}}{Smallest and biggest studied grain (resolution).}

\item{\code{aggregation}}{Pre-defined aggregation during running of \code{\link[=init_fmetrics]{init_fmetrics()}}.}

\item{\code{is_series}}{Logical. Whether the dataset was generated with direct databases from Global Forest Watch (FALSE) or from a cover time-series (TRUE).}

\item{\code{min_treecover}}{Pre-defined minimum tree cover during running of \code{\link[=init_fmetrics]{init_fmetrics()}}.}

\item{\code{min_cover}}{Pre-defined minimum forest cover during running of \code{\link[=init_fmetrics]{init_fmetrics()}}.}

\item{\code{min_rate}}{Pre-defined minimum forest loss rate during running of \code{\link[=init_fmetrics]{init_fmetrics()}}.}

\item{\code{window}}{Pre-defined window of years to calculate \code{min_rate} during running of \code{\link[=init_fmetrics]{init_fmetrics()}}.}

\item{\code{temporal_windows}}{Pre-defined time windows, relevant for activeness metric.}

\item{\code{excluded_cells}}{Cells excluded for future analysis (not frontiers)}
}}

