% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,init_FrontierMetric-method}
\alias{show,init_FrontierMetric-method}
\title{Shows 'init_FrontierMetric' object}
\usage{
\S4method{show}{init_FrontierMetric}(object)
}
\arguments{
\item{object}{Prints relevant information about a 'FrontierMetric' object.}
}
\value{
No return value.
}
\description{
Shows 'init_FrontierMetric' object
}
